+++ packages/rainbowkit/src/wallets/walletConnectors/index.ts
@@ -1,11 +1,13 @@
 import { argentWallet } from './argentWallet/argentWallet';
 import { bitKeepWallet } from './bitKeepWallet/bitKeepWallet';
+import { bitskiWallet } from './bitskiWallet/bitskiWallet';
 import { braveWallet } from './braveWallet/braveWallet';
 import { coinbaseWallet } from './coinbaseWallet/coinbaseWallet';
 import { imTokenWallet } from './imTokenWallet/imTokenWallet';
 import { injectedWallet } from './injectedWallet/injectedWallet';
 import { ledgerWallet } from './ledgerWallet/ledgerWallet';
 import { metaMaskWallet } from './metaMaskWallet/metaMaskWallet';
+import { mewWallet } from './mewWallet/mewWallet';
 import { omniWallet } from './omniWallet/omniWallet';
 import { rainbowWallet } from './rainbowWallet/rainbowWallet';
 import { trustWallet } from './trustWallet/trustWallet';
@@ -13,6 +15,7 @@ import { walletConnectWallet } from './walletConnectWallet/walletConnectWallet';
 
 export {
   argentWallet,
+  bitskiWallet,
   braveWallet,
   coinbaseWallet,
   imTokenWallet,
@@ -20,6 +23,7 @@ export {
   ledgerWallet,
   metaMaskWallet,
   bitKeepWallet,
+  mewWallet,
   omniWallet,
   rainbowWallet,
   trustWallet,